review-service/app/auth.py
from jose import jwt, JWTError
from fastapi import HTTPException, Depends, Request
from app.config import settings

def get_current_user(request: Request):
    token = request.headers.get("Authorization")
    if not token:
        raise HTTPException(status_code=401, detail="No token provided")
    try:
        scheme, _, token = token.partition(" ")
        payload = jwt.decode(token, settings.SECRET_KEY, algorithms=[settings.ALGORITHM])
        return payload.get("sub")
    except JWTError:
        raise HTTPException(status_code=403, detail="Invalid token")
