review-service/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
python-jose
python-dotenv
httpx
pydantic-settings